Convert to tyepscript.
As an experiment, I was trying to use "typesafe javascript". This approach uses jsdoc type decls + the IDE to achieve typesaftey w/out needing a compile step.

I gave it a go, but it was just wasn't as nice as actually using typescript. A compile step (webpack) also allows for some features like treeshaking for production... which is super nice because the codebase has a TON of debug-only code in it.

It should also make unit testing easier.